pyproject.toml
[project]
name = "vbp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "your@email.com"},
]
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[agent-engines]>=1.130.0,<1.147.0",
    "protobuf>=6.31.1,<7.0.0",
    "pymupdf>=1.27.2.2",
    "nltk>=3.9.4",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.3",
    "aiohttp>=3.9.5",
    "google-cloud-storage>=3.10.1",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-exporter-gcp-logging>=1.11.0a0",
    "jinja2>=3.1.0",
    "markdown>=3.6",
    "streamlit>=1.56.0",
    "pyrebase4>=4.9.0",
]
requires-python = ">=3.10,<3.15"


[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
    "pytest-cov>=7.1.0",
]

[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
]
eval = [
    "google-adk[eval]>=1.15.0,<2.0.0",
    "google-cloud-aiplatform[evaluation]>=1.130.0,<1.147.0",
]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "ty>=0.0.1a0",
    "codespell>=2.2.0,<3.0.0",
]
tools = [
    "jinja2>=3.1.0",
    "markdown>=3.6",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006", "E701", "E402", "E722", "F841"] # ignore line too long, too complex, multiple statements, module imports, bare except, unused variables

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ty]
# ty is Astral's Rust-based type checker (same team as ruff/uv)
# See: https://docs.astral.sh/ty/

[tool.ty.environment]
python-version = "3.14"

[tool.ty.src]
exclude = [".venv/**"]

[tool.ty.rules]
# Ignore common issues with third-party libraries and dynamic code patterns
unresolved-import = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
possibly-missing-attribute = "ignore"
not-subscriptable = "ignore"
deprecated = "ignore"

[tool.codespell]
ignore-words-list = "rouge,als,som,ned,mot,selv,fo,positiv,negativ,respons,rutine,alternativ,manuell,temperatur,tru,lokal,fase,generell,gir,noe,objekt"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb,**/package-lock.json,./tests/integration/results/*,*.pdf,**/*.json,**/*.csv,./local_data/*,./scripts/utils/results/*,./app/agents/clinical_taxonomist/data/*,./scripts/analysis/*"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.agent-starter-pack]
# Generation metadata - enables CLI commands to understand project context
# These fields allow any generated project to work as a remote template
name = "vbp"
description = ""
base_template = "local_VBP"
agent_directory = "app"

# Generation context
generated_at = "2026-04-07T18:41:32.245027+00:00"
asp_version = "0.40.1"

[tool.agent-starter-pack.create_params]
# Parameters used during project creation
deployment_target = "agent_engine"
session_type = "none"
cicd_runner = "skip"
include_data_ingestion = false
datastore = "none"

